fix(ci): fix cassandra test flake
