+++ metadata-io/src/test/java/com/linkedin/metadata/entity/cassandra/CassandraAspectMigrationsDaoTest.java
@@ -1,14 +1,15 @@
-package com.linkedin.metadata.timeline;
+package com.linkedin.metadata.entity.cassandra;
 
-import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.*;
 
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.linkedin.metadata.CassandraTestUtils;
 import com.linkedin.metadata.config.PreProcessHooks;
+import com.linkedin.metadata.entity.AspectMigrationsDaoTest;
 import com.linkedin.metadata.entity.EntityServiceImpl;
-import com.linkedin.metadata.entity.cassandra.CassandraAspectDao;
 import com.linkedin.metadata.event.EventProducer;
 import com.linkedin.metadata.models.registry.EntityRegistryException;
+import com.linkedin.metadata.service.UpdateIndicesService;
 import org.testcontainers.containers.CassandraContainer;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -16,19 +17,11 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-/**
- * A class that knows how to configure {@link TimelineServiceTest} to run integration tests against
- * a Cassandra database.
- *
- * <p>This class also contains all the test methods where realities of an underlying storage leak
- * into the {@link TimelineServiceImpl} in the form of subtle behavior differences. Ideally that
- * should never happen, and it'd be great to address captured differences.
- */
-public class CassandraTimelineServiceTest extends TimelineServiceTest<CassandraAspectDao> {
+public class CassandraAspectMigrationsDaoTest extends AspectMigrationsDaoTest<CassandraAspectDao> {
 
   private CassandraContainer _cassandraContainer;
 
-  public CassandraTimelineServiceTest() throws EntityRegistryException {}
+  public CassandraAspectMigrationsDaoTest() throws EntityRegistryException {}
 
   @BeforeClass
   public void setupContainer() {
@@ -48,15 +41,18 @@ public class CassandraTimelineServiceTest extends TimelineServiceTest<CassandraA
 
   private void configureComponents() {
     CqlSession session = CassandraTestUtils.createTestSession(_cassandraContainer);
-    _aspectDao = new CassandraAspectDao(session);
-    _aspectDao.setConnectionValidated(true);
-    _entityTimelineService = new TimelineServiceImpl(_aspectDao, _testEntityRegistry);
+    CassandraAspectDao dao = new CassandraAspectDao(session);
+    dao.setConnectionValidated(true);
     _mockProducer = mock(EventProducer.class);
+    _mockUpdateIndicesService = mock(UpdateIndicesService.class);
     PreProcessHooks preProcessHooks = new PreProcessHooks();
     preProcessHooks.setUiEnabled(true);
-    _entityServiceImpl =
-        new EntityServiceImpl(_aspectDao, _mockProducer, true, preProcessHooks, true);
+    _entityServiceImpl = new EntityServiceImpl(dao, _mockProducer, true, preProcessHooks, true);
     _entityServiceImpl.setUpdateIndicesService(_mockUpdateIndicesService);
+    _retentionService = new CassandraRetentionService(_entityServiceImpl, session, 1000);
+    _entityServiceImpl.setRetentionService(_retentionService);
+
+    _migrationsDao = dao;
   }
 
   /**
@@ -65,7 +61,7 @@ public class CassandraTimelineServiceTest extends TimelineServiceTest<CassandraA
    * test to make sure this class will always be discovered.
    */
   @Test
-  public void obligatoryTest() throws Exception {
+  public void obligatoryTest() throws AssertionError {
     Assert.assertTrue(true);
   }
 }
