+++ metadata-io/src/test/java/com/linkedin/metadata/timeline/cassandra/CassandraTimelineServiceTest.java
@@ -1,15 +1,16 @@
-package com.linkedin.metadata.entity;
+package com.linkedin.metadata.timeline.cassandra;
 
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.mock;
 
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.linkedin.metadata.CassandraTestUtils;
 import com.linkedin.metadata.config.PreProcessHooks;
+import com.linkedin.metadata.entity.EntityServiceImpl;
 import com.linkedin.metadata.entity.cassandra.CassandraAspectDao;
-import com.linkedin.metadata.entity.cassandra.CassandraRetentionService;
 import com.linkedin.metadata.event.EventProducer;
 import com.linkedin.metadata.models.registry.EntityRegistryException;
-import com.linkedin.metadata.service.UpdateIndicesService;
+import com.linkedin.metadata.timeline.TimelineServiceImpl;
+import com.linkedin.metadata.timeline.TimelineServiceTest;
 import org.testcontainers.containers.CassandraContainer;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -17,11 +18,19 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class CassandraAspectMigrationsDaoTest extends AspectMigrationsDaoTest<CassandraAspectDao> {
+/**
+ * A class that knows how to configure {@link TimelineServiceTest} to run integration tests against
+ * a Cassandra database.
+ *
+ * <p>This class also contains all the test methods where realities of an underlying storage leak
+ * into the {@link TimelineServiceImpl} in the form of subtle behavior differences. Ideally that
+ * should never happen, and it'd be great to address captured differences.
+ */
+public class CassandraTimelineServiceTest extends TimelineServiceTest<CassandraAspectDao> {
 
   private CassandraContainer _cassandraContainer;
 
-  public CassandraAspectMigrationsDaoTest() throws EntityRegistryException {}
+  public CassandraTimelineServiceTest() throws EntityRegistryException {}
 
   @BeforeClass
   public void setupContainer() {
@@ -41,18 +50,15 @@ public class CassandraAspectMigrationsDaoTest extends AspectMigrationsDaoTest<Ca
 
   private void configureComponents() {
     CqlSession session = CassandraTestUtils.createTestSession(_cassandraContainer);
-    CassandraAspectDao dao = new CassandraAspectDao(session);
-    dao.setConnectionValidated(true);
+    _aspectDao = new CassandraAspectDao(session);
+    _aspectDao.setConnectionValidated(true);
+    _entityTimelineService = new TimelineServiceImpl(_aspectDao, _testEntityRegistry);
     _mockProducer = mock(EventProducer.class);
-    _mockUpdateIndicesService = mock(UpdateIndicesService.class);
     PreProcessHooks preProcessHooks = new PreProcessHooks();
     preProcessHooks.setUiEnabled(true);
-    _entityServiceImpl = new EntityServiceImpl(dao, _mockProducer, true, preProcessHooks, true);
+    _entityServiceImpl =
+        new EntityServiceImpl(_aspectDao, _mockProducer, true, preProcessHooks, true);
     _entityServiceImpl.setUpdateIndicesService(_mockUpdateIndicesService);
-    _retentionService = new CassandraRetentionService(_entityServiceImpl, session, 1000);
-    _entityServiceImpl.setRetentionService(_retentionService);
-
-    _migrationsDao = dao;
   }
 
   /**
@@ -61,7 +67,7 @@ public class CassandraAspectMigrationsDaoTest extends AspectMigrationsDaoTest<Ca
    * test to make sure this class will always be discovered.
    */
   @Test
-  public void obligatoryTest() throws AssertionError {
+  public void obligatoryTest() throws Exception {
     Assert.assertTrue(true);
   }
 }
