+++ metadata-io/src/test/resources/testng-cassandra.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="metadata-io-cassandra" parallel="tests" thread-count="1">
+    <test name="cassandra-testcontainers">
+        <packages>
+            <package name="com.linkedin.metadata.entity.cassandra.*" />
+            <package name="com.linkedin.metadata.timeline.cassandra.*"/>
+        </packages>
+    </test>
+</suite>
